Optimise primitive_restart::upload_untouched()
This optimisation targets SSE4.1 and AVX2 and is only applied when
skip_restart is false.

I’ve only tested the u16 codepath, as it is the one used in NieR.

In some very unscientific profiling, this function used to take 2.76% of
the total time at some spot of the port town, it now takes about 0.40%
on SSE4.1 and 0.30% on AVX2.